requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-cloud-bigquery
numpy
cachetools
